requirements.txt
streamlit 
google-generativeai
python-dotenv 


Flask
Pillow
beautifulsoup4
bitarray
certifi
cffi
charset-normalizer
click
cryptography
dnspython
Flask
greenlet
idna
itsdangerous
Jinja2
markdown-it-py
MarkupSafe
mdurl
numpy
packaging
Pillow
pip
pycparser
pyparsing
python-dotenv
requests
setuptools
six
soupsieve
types-requests
types-setuptools
urllib3
Werkzeug
wheel